HelpEventDelegate: Move code to utils
